added component for dashboard metrics
